Settings adapted to Hostinger - Google Maps js added - SQL queries updated - Position class improved

+++ sql/create_positions.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS taller_2.positions (
   id INT NOT NULL AUTO_INCREMENT,
-  latitude INT NOT NULL,
-  longitude INT NOT NULL,
+  latitude FLOAT(9,6) NOT NULL,
+  longitude FLOAT(9,6) NOT NULL,
   registered_at DATETIME,
   PRIMARY KEY(id)
 );
